+++ src/sim/simulation/actions/buffer.py
@@ -1,4 +1,5 @@
 from logging import getLogger
+
 from simpy import Environment
 
 from src.sim.simulation.actions.action import Action
@@ -23,10 +24,12 @@ class Buffer(Action):
         super().__init__(env, warehouse, simulation)
 
     # override
-    def simulate_action(self):
-        simulation = self.simulation
-        carousel = self.warehouse.get_carousel()
-        env = self.env
+    def simulate_action(self, drawer=None, destination=None):
+        assert drawer is None, logger.warning("The buffer checks the two positions in the carousel, "
+                                              "so the drawer parameter is not taken into account")
+        assert destination is None, logger.warning("The default destination parameter is bay and buffer, "
+                                                   "so the destination parameter is not taken into account.")
+        simulation, carousel, env = self.simulation, self.warehouse.get_carousel(), self.env
         # try to take buffer resource
         with simulation.get_res_buffer().request() as req_buf:
             yield req_buf