+++ src/sim/simulation/actions/material/insert_material/insert_material.py
@@ -1,4 +1,3 @@
-from logging import getLogger
 from abc import abstractmethod
 
 from simpy import Environment
@@ -6,8 +5,6 @@ from simpy import Environment
 from src.sim.simulation.actions.action import Action
 from src.sim.simulation.simulation import Simulation
 from src.sim.warehouse import Warehouse
-
-logger = getLogger(__name__)
 
 
 class InsertMaterial(Action):
@@ -37,8 +34,6 @@ class InsertMaterial(Action):
         return self.duration
 
     @abstractmethod
-    def simulate_action(self):
-        logger.debug(f"Time {self.env.now:5.2f} - Start putting materials inside a drawer")
-        # return the drawer that is outside
-        return self.warehouse.get_carousel().get_deposit_entry().get_drawer()
+    def simulate_action(self, drawer=None, destination=None):
+        pass