+++ src/sim/simulation/actions/move/go_to_deposit.py
@@ -1,35 +1,34 @@
 from logging import getLogger
+
 from simpy import Environment
 
-from src.sim.drawer import Drawer
 from src.sim.simulation.actions.move.move import Move
 from src.sim.simulation.simulation import Simulation
-from src.sim.status_warehouse.enum_warehouse import EnumWarehouse
 from src.sim.warehouse import Warehouse
 
 logger = getLogger(__name__)
 
 
 class GoToDeposit(Move):
-    def __init__(self, env: Environment, warehouse: Warehouse, simulation: Simulation, drawer: Drawer,
-                 destination: EnumWarehouse):
+    def __init__(self, env: Environment, warehouse: Warehouse, simulation: Simulation):
         """
         Movement to go to the deposit.
 
         :type env: Environment
         :type warehouse: Warehouse
         :type simulation: Simulation
-        :type destination: EnumWarehouse
-        :type drawer: Drawer
         :param env: the simulation environment (SimPy Environment).
         :param warehouse: the warehouse where the action is performed.
         :param simulation: the simulation where the action is performed.
-        :param destination: the destination of the move.
-        :param drawer: the drawer used in the movement.
         """
-        super().__init__(env, warehouse, simulation, destination, drawer)
+        super().__init__(env, warehouse, simulation)
 
     # override
-    def simulate_action(self):
-        logger.debug(f"Time {self.env.now:5.2f} - Start come back to deposit position")
-        yield self.env.process(self.simulation.go_to_deposit())
+    def simulate_action(self, drawer=None, destination=None):
+        assert drawer is None, logger.warning("A go to deposit move is the default go to deposit move, "
+                                              "so the drawer parameter is not taken into account.")
+        assert destination is None, logger.warning("The default destination parameter is bay, "
+                                                   "so the destination parameter is not taken into account.")
+        env = self.env
+        logger.debug(f"Time {env.now:5.2f} - Start come back to deposit position")
+        yield env.process(self.simulation.go_to_deposit())