+++ src/sim/simulation/actions/move/vertical.py
@@ -1,36 +1,36 @@
 from logging import getLogger
+
 from simpy import Environment
 
 from src.sim.simulation.actions.move.move import Move
 from src.sim.simulation.simulation import Simulation
 from src.sim.status_warehouse.enum_warehouse import EnumWarehouse
-from src.sim.warehouse import Warehouse, Drawer
+from src.sim.warehouse import Warehouse
 
 logger = getLogger(__name__)
 
 
 class Vertical(Move):
-    def __init__(self, env: Environment, warehouse: Warehouse, simulation: Simulation, drawer: Drawer,
-                 destination: EnumWarehouse):
+    def __init__(self, env: Environment, warehouse: Warehouse, simulation: Simulation):
         """
         Vertical movement.
 
         :type env: Environment
         :type warehouse: Warehouse
         :type simulation: Simulation
-        :type destination: EnumWarehouse
-        :type drawer: Drawer
         :param env: the simulation environment (SimPy Environment).
         :param warehouse: the warehouse where the action is performed.
         :param simulation: the simulation where the action is performed.
-        :param destination: the destination of the move.
-        :param drawer: the drawer used in the movement.
         """
-        super().__init__(env, warehouse, simulation, destination, drawer)
+        super().__init__(env, warehouse, simulation)
 
-    def simulate_action(self):
-        logger.debug(f"Time {self.env.now:5.2f} - Start vertical move")
-        if self.get_destination() == EnumWarehouse.CAROUSEL:
-            yield self.env.process(self.simulation.reach_drawer_height(self.drawer))
-        else:
-            yield self.env.process(self.simulation.allocate_best_pos(self.drawer))
+    def simulate_action(self, drawer=None, destination=None):
+        assert drawer is not None, logger.error("The drawer cannot be None!")
+        assert destination is not None, logger.error("The destination cannot be None!")
+        env = self.env
+        logger.debug(f"Time {env.now:5.2f} - Start vertical move")
+        yield env.process(
+            self.simulation.reach_drawer_height(drawer)
+            if destination == EnumWarehouse.CAROUSEL else
+            self.simulation.allocate_best_pos(drawer)
+        )